requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
distlib==0.4.0
dnspython==2.7.0
elastic-transport==8.11.0
elasticsearch==8.11.0
email-validator==2.3.0
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
filelock==3.20.0
frozenlist==1.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.11
importlib_metadata==6.11.0
iniconfig==2.3.0
Jinja2==3.1.6
loguru==0.7.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
nodeenv==1.9.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-fastapi-instrumentator==7.0.0
python-logging-loki==0.3.1
packaging==25.0
passlib==1.7.4
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.5.0
propcache==0.4.1
protobuf==4.25.5
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
ruff==0.14.7
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
slugify==0.0.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
stripe==14.0.1
text-unidecode==1.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
watchfiles==1.1.1
websockets==15.0.1
win32_setctime==1.2.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
